consumer/requirements.txt
kafka-python==2.0.2
boto3
pandas==2.0.3